All features regained...
Toggling a switch, should toggle an output with a matching ID.
This time PCFs are used asynchronously, previously they were blocking, and
especially reading IOs was slow - separately reading the PCF for EACH single
PIN. Instead of doing a single call. This is much better now.

NOT tested with hardware.